WIP: refactor SearchService
